did a ton of refactoring (collisions/grav/leveleditor/animations) (read description)
1- polished the gravity code so it finally works exactly how we
expected.
> we still need to make the game feel actually fun, but at least now
there is no bugs in the code for gravity and collisions. what we are
missing now is figure out how to make the movement feel fun, and change
the way gravity works to make it feel more like a fluid game.
2- also fixed the animation system finally.
> hard coded a lot of stuff in player_input so we got the expected
behavior out of the animations. the idea is to move all of that out of
the player system into an event class thing so we can add more stuff
later easier. and also the code is insanely coupled and it sucks.
3- fixed the level editor to actually work with the new code
4- added a sample started level with 2 rectangles spawned from main